Move kas_core::event::config -> kas_core::config::event
