+++ crates/kas-core/src/app/window.rs
@@ -9,8 +9,9 @@ use super::common::WindowSurface;
 use super::shared::{AppSharedState, AppState};
 use super::{AppData, AppGraphicsBuilder, ProxyAction};
 use crate::cast::{Cast, Conv};
+use crate::config::event::WindowConfig;
 use crate::draw::{color::Rgba, AnimationState, DrawSharedImpl};
-use crate::event::{config::WindowConfig, ConfigCx, CursorIcon, EventState};
+use crate::event::{ConfigCx, CursorIcon, EventState};
 use crate::geom::{Coord, Rect, Size};
 use crate::layout::SolveCache;
 use crate::theme::{DrawCx, SizeCx, Theme, ThemeSize, Window as _};
